Move test_subclass_before_meta_init() to test_meta.py
