+++ tests/test_meta.py
@@ -1,7 +1,11 @@
+import logging
+
 import pytest
 
 from fonduer import Meta
+from fonduer.candidates.models import mention_subclass
 
+logger = logging.getLogger(__name__)
 DB = "meta_test"
 
 
@@ -21,3 +25,14 @@ def test_meta_connection_strings():
     assert Meta.DBNAME == DB
     Meta.init("postgresql://localhost:5432/" + "cand_test").Session()
     assert Meta.DBNAME == "cand_test"
+
+
+def test_subclass_before_meta_init():
+    """Test if it is possible to create a mention (candidate) subclass even before Meta
+    is initialized.
+    """
+    Part = mention_subclass("Part")
+    logger.info(f"Create a mention subclass '{Part.__tablename__}'")
+    Meta.init("postgresql://localhost:5432/" + DB).Session()
+    Temp = mention_subclass("Temp")
+    logger.info(f"Create a mention subclass '{Temp.__tablename__}'")
